requirements.txt
absl-py==1.0.0
aiohttp==3.9.5
aiosignal==1.3.1
appdirs==1.4.4
appnope==0.1.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
asttokens==2.0.5
async-timeout==4.0.3
attrs==23.1.0
audioread==2.1.9
backcall==0.2.0
backports.functools-lru-cache==1.6.4
beautifulsoup4==4.11.1
bleach==4.1.0
cachetools==5.0.0
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==2.0.12
click==8.1.7
cycler==0.11.0
datasets==2.20.0
debugpy==1.6.0
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.8
docopt==0.6.2
entrypoints==0.4
evaluate==0.4.2
executing==0.8.3
fastjsonschema==2.16.2
filelock==3.6.0
fonttools==4.32.0
frozenlist==1.4.1
fsspec==2024.5.0
gdown==4.4.0
gensim==4.3.2
google-api-core==2.19.0
google-api-python-client==2.134.0
google-auth==2.30.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==0.4.6
googleapis-common-protos==1.63.1
graphviz==0.20
grpcio==1.44.0
httplib2==0.22.0
huggingface-hub==0.23.4
idna==3.3
importlib-metadata==7.0.1
importlib-resources==6.1.1
inflection==0.5.1
ipykernel==6.12.1
ipython==8.12.2
ipython-genutils==0.2.0
ipywidgets==7.7.0
jedi==0.18.1
Jinja2==3.1.4
joblib==1.1.0
jsonschema==4.19.2
jsonschema-specifications==2023.7.1
jupyter-client==7.2.2
jupyter_core==5.7.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==1.0.0
kiwisolver==1.4.2
librosa==0.8.1
llvmlite==0.38.0
Markdown==3.3.6
MarkupSafe==2.1.3
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==2.0.4
more-itertools==8.12.0
multidict==6.0.5
multiprocess==0.70.16
nbclient==0.10.0
nbconvert==7.10.0
nbformat==5.10.4
nest-asyncio==1.6.0
nltk==3.8.1
notebook==6.4.8
numba==0.55.1
numpy==1.21.0
oauthlib==3.2.0
packaging==23.2
pandas==1.4.2
pandocfilters==1.5.0
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.1.0
pip==24.0
pipreqs==0.5.0
pkgutil_resolve_name==1.3.10
platformdirs==3.10.0
pooch==1.6.0
prometheus-client==0.14.1
prompt-toolkit==3.0.43
proto-plus==1.24.0
protobuf==4.25.3
psutil==5.9.0
ptyprocess==0.7.0
pure-eval==0.2.2
pyaml==24.4.0
pyarrow==16.1.0
pyarrow-hotfix==0.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydotplus==2.0.2
Pygments==2.15.1
pyparsing==3.0.7
pyrsistent==0.18.1
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2022.1
PyYAML==6.0.1
pyzmq==25.1.2
Quandl==3.7.0
referencing==0.30.2
regex==2023.12.25
requests==2.32.3
requests-oauthlib==1.3.1
resampy==0.2.2
rpds-py==0.10.6
rsa==4.8
scikit-learn==1.0.2
scipy==1.8.0
seaborn==0.11.2
Send2Trash==1.8.2
setuptools==69.5.1
six==1.16.0
smart-open==7.0.1
SoundFile==0.10.3.post1
soupsieve==2.3.2
stack-data==0.2.0
tensorboard==2.8.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
terminado==0.17.1
testpath==0.6.0
threadpoolctl==3.1.0
tinycss2==1.2.1
tokenizers==0.13.3
torchaudio==0.11.0
torchsummary==1.5.1
torchvision==0.12.0
torchviz==0.0.2
tornado==6.4.1
tqdm==4.66.4
traitlets==5.14.3
transformers==4.23.0.dev0
typing_extensions==4.11.0
uritemplate==4.1.1
urllib3==1.26.9
wcwidth==0.2.5
webencodings==0.5.1
Werkzeug==2.1.1
wheel==0.43.0
widgetsnbextension==3.6.0
wrapt==1.16.0
xxhash==3.4.1
yarg==0.1.9
yarl==1.9.4
zipp==3.17.0
